requirements.txt
biopython>=1.80
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
requests>=2.31
